feature:Remove zk3.4 support in 3.3 (#13960)
