+++ .artifacts
@@ -39,7 +39,6 @@ dubbo-dependencies
 dubbo-dependencies-all
 dubbo-dependencies-bom
 dubbo-dependencies-zookeeper
-dubbo-dependencies-zookeeper-curator5
 dubbo-distribution
 dubbo-filter-cache
 dubbo-filter-validation
@@ -82,7 +81,6 @@ dubbo-remoting-netty
 dubbo-remoting-netty4
 dubbo-remoting-zookeeper-api
 dubbo-remoting-zookeeper
-dubbo-remoting-zookeeper-curator5
 dubbo-rpc
 dubbo-rpc-api
 dubbo-rpc-dubbo
