+++ dubbo-config/dubbo-config-spring/pom.xml
@@ -194,7 +194,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-remoting-zookeeper-curator5</artifactId>
+      <artifactId>dubbo-remoting-zookeeper</artifactId>
       <version>${project.parent.version}</version>
       <scope>test</scope>
     </dependency>
