+++ dubbo-dependencies/pom.xml
@@ -28,7 +28,6 @@
   <packaging>pom</packaging>
   <modules>
     <module>dubbo-dependencies-zookeeper</module>
-    <module>dubbo-dependencies-zookeeper-curator5</module>
   </modules>
 
   <properties>
