+++ dubbo-registry/dubbo-registry-zookeeper/pom.xml
@@ -38,7 +38,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-remoting-zookeeper-curator5</artifactId>
+      <artifactId>dubbo-remoting-zookeeper</artifactId>
       <version>${project.parent.version}</version>
     </dependency>
     <dependency>
