--- dubbo-remoting/dubbo-remoting-zookeeper-curator5/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.remoting.zookeeper.ZookeeperTransporter
@@ -1,1 +1,0 @@
-curator5=org.apache.dubbo.remoting.zookeeper.curator5.Curator5ZookeeperTransporter
