+++ dubbo-remoting/dubbo-remoting-zookeeper/pom.xml
@@ -29,9 +29,10 @@
   <description>The zookeeper remoting module of dubbo project</description>
   <properties>
     <skip_maven_deploy>false</skip_maven_deploy>
-    <zookeeper_version>3.4.14</zookeeper_version>
-    <curator_version>4.3.0</curator_version>
+    <curator5_version>5.1.0</curator5_version>
+    <zookeeper_version>3.7.2</zookeeper_version>
   </properties>
+
   <dependencies>
     <dependency>
       <groupId>org.apache.dubbo</groupId>
@@ -45,26 +46,14 @@
     </dependency>
     <dependency>
       <groupId>org.apache.curator</groupId>
-      <artifactId>curator-recipes</artifactId>
-      <version>${curator_version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.curator</groupId>
       <artifactId>curator-framework</artifactId>
-      <version>${curator_version}</version>
+      <version>${curator5_version}</version>
     </dependency>
     <dependency>
       <groupId>org.apache.curator</groupId>
-      <artifactId>curator-x-discovery</artifactId>
-      <version>${curator_version}</version>
+      <artifactId>curator-recipes</artifactId>
+      <version>${curator5_version}</version>
     </dependency>
-
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-native</artifactId>
-      <version>${project.parent.version}</version>
-    </dependency>
-
     <dependency>
       <groupId>org.apache.zookeeper</groupId>
       <artifactId>zookeeper</artifactId>
@@ -73,8 +62,12 @@
 
     <dependency>
       <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-test-common</artifactId>
+      <artifactId>dubbo-native</artifactId>
       <version>${project.parent.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.logging.log4j</groupId>
+      <artifactId>log4j-slf4j-impl</artifactId>
       <scope>test</scope>
     </dependency>
   </dependencies>
