+++ dubbo-remoting/dubbo-remoting-zookeeper/src/main/java/org/apache/dubbo/remoting/zookeeper/curator/CuratorZookeeperClient.java
@@ -20,8 +20,6 @@ import org.apache.dubbo.common.URL;
 import org.apache.dubbo.common.config.configcenter.ConfigItem;
 import org.apache.dubbo.common.logger.ErrorTypeAwareLogger;
 import org.apache.dubbo.common.logger.LoggerFactory;
-import org.apache.dubbo.common.utils.NamedThreadFactory;
-import org.apache.dubbo.common.utils.StringUtils;
 import org.apache.dubbo.remoting.zookeeper.AbstractZookeeperClient;
 import org.apache.dubbo.remoting.zookeeper.ChildListener;
 import org.apache.dubbo.remoting.zookeeper.DataListener;
@@ -32,11 +30,8 @@ import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executor;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.curator.framework.CuratorFramework;
@@ -72,7 +67,7 @@ public class CuratorZookeeperClient
 
     private static final Charset CHARSET = StandardCharsets.UTF_8;
     private final CuratorFramework client;
-    private static final Map<String, NodeCache> nodeCacheMap = new ConcurrentHashMap<>();
+    private static Map<String, NodeCache> nodeCacheMap = new ConcurrentHashMap<>();
 
     public CuratorZookeeperClient(URL url) {
         super(url);
@@ -85,7 +80,7 @@ public class CuratorZookeeperClient
                     .connectionTimeoutMs(timeout)
                     .sessionTimeoutMs(sessionExpireMs);
             String userInformation = url.getUserInformation();
-            if (StringUtils.isNotEmpty(userInformation)) {
+            if (userInformation != null && userInformation.length() > 0) {
                 builder = builder.authorization("digest", userInformation.getBytes());
                 builder.aclProvider(new ACLProvider() {
                     @Override
@@ -119,9 +114,7 @@ public class CuratorZookeeperClient
                 throw illegalStateException;
             }
 
-            CuratorWatcherImpl.closed = false;
         } catch (Exception e) {
-            close();
             throw new IllegalStateException(e.getMessage(), e);
         }
     }
@@ -347,15 +340,8 @@ public class CuratorZookeeperClient
 
     @Override
     public void doClose() {
-        super.close();
+        super.doClose();
         client.close();
-        CuratorWatcherImpl.closed = true;
-        synchronized (CuratorWatcherImpl.class) {
-            if (CuratorWatcherImpl.CURATOR_WATCHER_EXECUTOR_SERVICE != null) {
-                CuratorWatcherImpl.CURATOR_WATCHER_EXECUTOR_SERVICE.shutdown();
-                CuratorWatcherImpl.CURATOR_WATCHER_EXECUTOR_SERVICE = null;
-            }
-        }
     }
 
     @Override
@@ -377,7 +363,7 @@ public class CuratorZookeeperClient
     @Override
     protected CuratorZookeeperClient.NodeCacheListenerImpl createTargetDataListener(
             String path, DataListener listener) {
-        return new NodeCacheListenerImpl(listener, path);
+        return new NodeCacheListenerImpl(client, listener, path);
     }
 
     @Override
@@ -422,13 +408,16 @@ public class CuratorZookeeperClient
 
     static class NodeCacheListenerImpl implements NodeCacheListener {
 
+        private CuratorFramework client;
+
         private volatile DataListener dataListener;
 
         private String path;
 
         protected NodeCacheListenerImpl() {}
 
-        public NodeCacheListenerImpl(DataListener dataListener, String path) {
+        public NodeCacheListenerImpl(CuratorFramework client, DataListener dataListener, String path) {
+            this.client = client;
             this.dataListener = dataListener;
             this.path = path;
         }
@@ -453,24 +442,9 @@ public class CuratorZookeeperClient
 
     static class CuratorWatcherImpl implements CuratorWatcher {
 
-        private static volatile ExecutorService CURATOR_WATCHER_EXECUTOR_SERVICE;
-
-        private static volatile boolean closed = false;
-
         private CuratorFramework client;
         private volatile ChildListener childListener;
         private String path;
-
-        private static void initExecutorIfNecessary() {
-            if (!closed && CURATOR_WATCHER_EXECUTOR_SERVICE == null) {
-                synchronized (CuratorWatcherImpl.class) {
-                    if (!closed && CURATOR_WATCHER_EXECUTOR_SERVICE == null) {
-                        CURATOR_WATCHER_EXECUTOR_SERVICE =
-                                Executors.newSingleThreadExecutor(new NamedThreadFactory("Dubbo-CuratorWatcher"));
-                    }
-                }
-            }
-        }
 
         public CuratorWatcherImpl(CuratorFramework client, ChildListener listener, String path) {
             this.client = client;
@@ -493,27 +467,15 @@ public class CuratorZookeeperClient
             }
 
             if (childListener != null) {
-                Runnable task = () -> Optional.ofNullable(childListener).ifPresent(c -> {
-                    try {
-                        c.childChanged(
-                                path,
-                                client.getChildren()
-                                        .usingWatcher(CuratorWatcherImpl.this)
-                                        .forPath(path));
-                    } catch (Exception e) {
-                        logger.warn(REGISTRY_ZOOKEEPER_EXCEPTION, "", "", "client get children error", e);
-                    }
-                });
-                initExecutorIfNecessary();
-                if (!closed && CURATOR_WATCHER_EXECUTOR_SERVICE != null) {
-                    CURATOR_WATCHER_EXECUTOR_SERVICE.execute(task);
-                }
+                childListener.childChanged(
+                        path, client.getChildren().usingWatcher(this).forPath(path));
             }
         }
     }
 
     private class CuratorConnectionStateListener implements ConnectionStateListener {
         private final long UNKNOWN_SESSION_ID = -1L;
+
         private long lastSessionId;
         private int timeout;
         private int sessionExpireMs;
