+++ dubbo-remoting/dubbo-remoting-zookeeper/src/test/java/org/apache/dubbo/remoting/zookeeper/curator/CuratorZookeeperTransporterTest.java
@@ -22,14 +22,11 @@ import org.apache.dubbo.remoting.zookeeper.ZookeeperClient;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.DisabledForJreRange;
-import org.junit.jupiter.api.condition.JRE;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsNot.not;
 import static org.hamcrest.core.IsNull.nullValue;
 
-@DisabledForJreRange(min = JRE.JAVA_16)
 class CuratorZookeeperTransporterTest {
     private ZookeeperClient zookeeperClient;
     private CuratorZookeeperTransporter curatorZookeeperTransporter;
@@ -41,7 +38,7 @@ class CuratorZookeeperTransporterTest {
     }
 
     @BeforeEach
-    public void setUp() throws Exception {
+    public void setUp() {
         zookeeperClient =
                 new CuratorZookeeperTransporter().connect(URL.valueOf(zookeeperConnectionAddress1 + "/service"));
         curatorZookeeperTransporter = new CuratorZookeeperTransporter();
