--- dubbo-remoting/dubbo-remoting-zookeeper/src/test/resources/dubbo.properties
@@ -1,2 +1,0 @@
-dubbo.application.enable-file-cache=false
-dubbo.service.shutdown.wait=200
