+++ dubbo-remoting/pom.xml
@@ -33,7 +33,6 @@
     <module>dubbo-remoting-netty</module>
     <module>dubbo-remoting-zookeeper-api</module>
     <module>dubbo-remoting-zookeeper</module>
-    <module>dubbo-remoting-zookeeper-curator5</module>
     <module>dubbo-remoting-netty4</module>
     <module>dubbo-remoting-http12</module>
   </modules>
