+++ dubbo-test/dubbo-dependencies-all/pom.xml
@@ -330,11 +330,6 @@
       <artifactId>dubbo-remoting-zookeeper</artifactId>
       <version>${project.version}</version>
     </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-remoting-zookeeper-curator5</artifactId>
-      <version>${project.version}</version>
-    </dependency>
 
     <!-- rpc -->
     <dependency>
